# Define attributes for score best model